refactor(routes): extract /file-server into file_server.js (god-file decomposition)
Route contract unchanged (284 routes), 0 /file-server left inline, vitest 38 passed 268 passed . Lucy (GLM-5.2), driver-gated.
